lb: update
