+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/urnetwork/warp
+
+go 1.24.0
