+++ lb/Makefile
@@ -1,11 +1,21 @@
 
-all: clean build image
+all: init clean build
 
 clean:
 	rm -rf build
 
+build:
+	env GOOS=linux GOARCH=arm64 go build -ldflags "-X main.Version=${WARP_VERSION}" -o build/linux/arm64/
+	env GOOS=linux GOARCH=amd64 go build -ldflags "-X main.Version=${WARP_VERSION}" -o build/linux/amd64/
+	env GOOS=darwin GOARCH=arm64 go build -ldflags "-X main.Version=${WARP_VERSION}" -o build/darwin/arm64/
+	env GOOS=darwin GOARCH=amd64 go build -ldflags "-X main.Version=${WARP_VERSION}" -o build/darwin/amd64/
+
+init:
+	go clean -cache
+	go clean -modcache
+
 warp_build:
-	$(MAKE) clean
+	$(MAKE) all
 	# generate the config
 	mkdir -p build/${WARP_ENV}/nginx.conf
 	warpctl lb create-config ${WARP_ENV} --out=build/${WARP_ENV}/nginx.conf;
