+++ lb/go.mod
@@ -1,0 +1,7 @@
+module github.com/urnetwork/warp/lb
+
+go 1.24.0
+
+require github.com/urnetwork/warp v0.0.0
+
+replace github.com/urnetwork/warp => ..
