--- warpctl/audit_controller.go
@@ -1,45 +1,0 @@
-package main
-
-func announceBuild(env string, service string, version string) {
-	// post to discord
-}
-
-func announceDeployStarted(env string, service string, deployBlocks []string, deployVersion string) {
-	// post to discord
-}
-
-func announceDeployEnded(env string, service string, deployBlocks []string, deployVersion string) {
-	// post to discord
-}
-
-func announceRunWaitForVersion() {
-
-}
-
-func announceRunWaitForConfig() {
-
-}
-
-func announceRunStart() {
-
-}
-
-func announceRunError() {
-
-}
-
-func announceRunFail() {
-
-}
-
-func announceRunSuccess() {
-
-}
-
-func announceRunEnter() {
-
-}
-
-func announceRunExit() {
-
-}
