--- warpctl/events.go
@@ -1,63 +1,0 @@
-package main
-
-import (
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
-)
-
-type Event struct {
-	set chan bool
-}
-
-func NewEvent() *Event {
-	return &Event{
-		set: make(chan bool, 0),
-	}
-}
-
-func (self *Event) Set() {
-	close(self.set)
-}
-
-func (self *Event) IsSet() bool {
-	select {
-	case <-self.set:
-		return true
-	default:
-		return false
-	}
-}
-
-func (self *Event) WaitForSet(timeout time.Duration) bool {
-	select {
-	case <-self.set:
-		return true
-	case <-time.After(timeout):
-		return false
-	}
-}
-
-func (self *Event) SetOnSignals(signalValues ...syscall.Signal) func() {
-	stopSignal := make(chan os.Signal, len(signalValues))
-	for _, signalValue := range signalValues {
-		signal.Notify(stopSignal, signalValue)
-	}
-	go func() {
-		for {
-			select {
-			case _, ok := <-stopSignal:
-				if ok {
-					self.Set()
-				} else {
-					return
-				}
-			}
-		}
-	}()
-	return func() {
-		signal.Stop(stopSignal)
-		close(stopSignal)
-	}
-}
